finalized exit key
